Add pipeline for managing the postgres-cluster bosh release (#142)
Introduced default paramaters (default_parameters.yml) when running the
pipelines. These parameters contains values used on our main concourse
intallation and can be overriden in the credentials.yml file (generate
script updated also to account for these. The set_pipeline.sh script was
updated to load the content of that new file.

This allows us to use different values for some properties on the AVA
environment specifically in this commit it will allow us to use Github
releases and on AVA it will just create draft releases with a different
name.

Fixes: VITE-831